refactored an analyzer to analyzer package
